+++ cogs/moderation.py
@@ -1,3 +1,4 @@
+import discord
 from discord.ext import commands
 from helpers import timestamp
 
@@ -15,5 +16,28 @@ class Moderation(commands.Cog):
             print(f'{ctx.message.author} deleted {amount} messages.')
             timestamp()
             
+    @commands.command()
+    async def joined(self, ctx: commands.Context, member: discord.Member):
+        joined = (f'{member.name} joined on {discord.utils.format_dt(member.joined_at)}.')
+        print(f'{joined}')
+        await ctx.send(f'{joined}')
+        timestamp()
+        
+    @commands.command
+    async def say(self, ctx: commands.Context, message: str):
+        print(f"{ctx.message.author} made McSwitch say:")
+        print(f'{message}')
+        timestamp()
+        await ctx.channel.purge(limit=1)
+        await ctx.send(f'{message}')
+        
+    @commands.command
+    async def playing(self, ctx: commands.Context, game: str, field: str, value: any):
+        embedPlaying = discord.Embed(title=game, color=purple)
+        embedPlaying.add_field(name=f'{field}', value=f'{value}', inline=True)
+        print(f'{ctx.message.author} is playing {game}: {field}, {value}')
+        timestamp()
+        await ctx.send(embed=embedPlaying)
+
 async def setup(bot: commands.Bot) -> None:
     await bot.add_cog(Moderation(bot))