+++ helpers.py
@@ -1,5 +1,9 @@
 from datetime import datetime
 import subprocess
+from datetime import datetime
+import discord
+
+sTime = datetime.now()
 
 def mcswitch():
     result = subprocess.run(["sh", "./mc.sh"], shell=True, capture_output=True, text=True)
@@ -21,3 +25,7 @@ def timestamp():
     cTime = datetime.now()
     print(f'{cTime}')
     print(f'----------')
+    
+async def send_dm(ctx, member: discord.Member, *, content):
+    channel = await member.create_dm()
+    await channel.send(content)